Split types.ts by domain; type builder functions; restore PORT-NOTEs
Extract path.ts (Path + path helpers), flags.ts (Flag/valFlag consts +
flagUnsafeHas), and tags.ts (Tag + tag/tagFlag consts + tagFlags) out of
the 505-line types.ts grab-bag, so each module imports from the domain it
needs instead of one bag.

Move the Builder and Encoder codegen-function types into builder.ts (their
natural home) and apply Builder/Encoder consistently to every builder
function that was still bare (unionDecoder, shapedParser, shapedSerializer,
compactColumnsDecoder, noopDecoder). types.ts and schema.ts reference them
via type-only upward imports, which the acyclic layering permits.

Restore the PORT-NOTE comments that document non-obvious ReScript->TS
translation decisions (reserved-word renames, @unboxed representations,
arguments/function constraints, no-op re-typings). Migration scaffolding
(TODO(integration), Fragment banners) stays removed.

Codegen is byte-identical; the tuple2 golden even shrank (9890->9788) from
the shifted tree-shaking boundaries.
