+++ packages/sury/specs/tuple2.yaml
@@ -4,7 +4,7 @@ ts:
   input: "[string, number]"
   output: "[string, number]"
   instantiations: 1168
-  bundleBytes: 9890
+  bundleBytes: 9788
 jsonSchema:
   input: '{ type: "array", minItems: 2, maxItems: 2, items: [{ type: "string" }, { type: "number" }] }'
   output: '{ type: "array", minItems: 2, maxItems: 2, items: [{ type: "string" }, { type: "number" }] }'
