+++ packages/sury/src/composites.ts
@@ -1,8 +1,14 @@
 import { Literal_parse, isArrayCond, jsonName, objectTagCond, setHas, unit } from "./primitives";
 import { baseSchema, getOrRethrow, panic, reversedKey, unknown, updateOutput } from "./schema";
 import { getOutputSchema, nestedLoc, nestedOptionParser, never_, parse, parseDynamic, typeCheckCond } from "./parse";
-import { B_Val_Object_add, B_Val_addKey, B_Val_scope, B_asyncVal, B_dynamicScope, B_embed, B_failWithArg, B_hoistChildChecks, B_hoistDecl, B_inlineConst, B_inlineLocation, B_isHoistable, B_makeInvalidInputDetails, B_markOutput, B_merge, B_mergeWithPathPrepend, B_next, B_nextConst, B_pushCheck, B_refine, B_throw, B_unsupportedDecode, B_varWithoutAllocation, _notVar, _notVarAtParent, _var, failInvalidType } from "./builder";
-import { Check, ErrorDetails, Internal, SuryErrorRecord, Val, arrayTag, flagUnsafeHas, immutableEmptyArray, immutableEmptyObject, isLiteral, isOptional, nullTag, numberTag, objectTag, pathConcat, pathFromInlinedLocation, tagFlagArray, tagFlagFunction, tagFlagInstance, tagFlagNaN, tagFlagNever, tagFlagNull, tagFlagObject, tagFlagRef, tagFlagUndefined, tagFlagUnion, tagFlagUnknown, tagFlags, undefinedTag, unionTag, unknownTag, valFlagAsync, valFlagNone } from "./types";
+import { B_Val_Object_add, B_Val_addKey, B_Val_scope, B_asyncVal, B_dynamicScope, B_embed, B_failWithArg, B_hoistChildChecks, B_hoistDecl, B_inlineConst, B_inlineLocation, B_isHoistable, B_makeInvalidInputDetails, B_markOutput, B_merge, B_mergeWithPathPrepend, B_next, B_nextConst, B_pushCheck, B_refine, B_throw, B_unsupportedDecode, B_varWithoutAllocation, Builder, _notVar, _notVarAtParent, _var, failInvalidType } from "./builder";
+import { Check, ErrorDetails, Internal, SuryErrorRecord, Val, immutableEmptyArray, immutableEmptyObject, isLiteral, isOptional } from "./types";
+import { flagUnsafeHas, valFlagAsync, valFlagNone } from "./flags";
+import { pathConcat, pathFromInlinedLocation } from "./path";
+import { arrayTag, nullTag, numberTag, objectTag, tagFlagArray, tagFlagFunction, tagFlagInstance, tagFlagNaN, tagFlagNever, tagFlagNull, tagFlagObject, tagFlagRef, tagFlagUndefined, tagFlagUnion, tagFlagUnknown, tagFlags, undefinedTag, unionTag, unknownTag } from "./tags";
+
+// PORT-NOTE: `B_Val_Object_t` is `{...val}` — the same runtime shape as `val`,
+// so this port uses the prelude's `Val` type for object vals.
 
 export const makeObjectVal = (prev: Val, schema: Internal): Val => {
   return {
@@ -604,7 +610,7 @@ export const unionEncoder = (input: Val, target: Internal): Val => {
   }
 }
 
-export const unionDecoder = (input: Val): Val => {
+export const unionDecoder: Builder = (input: Val) => {
   const selfSchema = input.e;
   let schemas = selfSchema.anyOf!;
   const initialInputTagFlag = tagFlags[input.s.type]!;
@@ -670,6 +676,9 @@ export const unionDecoder = (input: Val): Val => {
       return `${B_embed(
         input,
         // Reads `arguments`, so this must stay a `function` expression.
+        // PORT-NOTE: the source lambda reads `arguments`, so this must stay a
+        // `function` expression (X.Function.toExpression made it a plain
+        // uncurried function in ReScript; a TS function expression already is).
         function () {
           const args = arguments;
           B_throw(
@@ -718,6 +727,9 @@ export const unionDecoder = (input: Val): Val => {
       // If we come across an item without a discriminant
       // and without any code, it means that this item is always valid
       // and we should exit early
+      // PORT-NOTE: `itemCode = Single(string) | Multiple(array<string>)` is
+      // @unboxed — runtime value is the string itself or the array itself, so
+      // the cases are discriminated with Array.isArray.
       let byDiscriminant: Record<string, string | string[]> = {};
 
       const preItems = 2;
@@ -1294,6 +1306,8 @@ export const nestedOption = (item: Internal): Internal => {
   });
 }
 
+// PORT-NOTE: the `~unit` labeled arg is renamed to `unitSchema` so the
+// default expression can still reference the module-level `unit` factory.
 export const optionFactory = (item: Internal, unitSchema: Internal = unit()): Internal => {
   const out = getOutputSchema(item);
   if (out.type === undefinedTag) {
