+++ packages/sury/src/primitives.ts
@@ -1,6 +1,9 @@
 import { baseSchema, cached } from "./schema";
 import { B_embed, B_embedInvalidInput, B_inlineConst, B_next, B_nextConst, B_refine, B_unsupportedDecode, B_varWithoutAllocation, _var, failInvalidType } from "./builder";
-import { Builder, Check, Internal, Tag, Val, bigintTag, booleanTag, flagDisableNanNumberValidation, flagUnsafeHas, instanceTag, isLiteral, nanTag, nullTag, numberTag, objectTag, stringTag, symbolTag, tagFlagBigint, tagFlagBoolean, tagFlagNaN, tagFlagNull, tagFlagNumber, tagFlagRef, tagFlagString, tagFlagSymbol, tagFlagUndefined, tagFlagUnion, tagFlagUnknown, tagFlags, undefinedTag, unknownTag } from "./types";
+import { Check, Internal, Val, isLiteral } from "./types";
+import { Builder } from "./builder";
+import { flagDisableNanNumberValidation, flagUnsafeHas } from "./flags";
+import { Tag, bigintTag, booleanTag, instanceTag, nanTag, nullTag, numberTag, objectTag, stringTag, symbolTag, tagFlagBigint, tagFlagBoolean, tagFlagNaN, tagFlagNull, tagFlagNumber, tagFlagRef, tagFlagString, tagFlagSymbol, tagFlagUndefined, tagFlagUnion, tagFlagUnknown, tagFlags, undefinedTag, unknownTag } from "./tags";
 
 export const int32FormatValidation = (inputVar: string) => {
   return `${inputVar}<=2147483647&&${inputVar}>=-2147483648&&${inputVar}%1===0`;
@@ -14,6 +17,8 @@ export const nanCond = (inputVar: string): string => `Number.isNaN(${inputVar})`
 export const isArrayCond = (inputVar: string): string => `Array.isArray(${inputVar})`;
 export const objectTagCond = (inputVar: string): string =>
   `${typeofCond(objectTag)(inputVar)}&&${inputVar}`;
+// PORT-NOTE: `class` is a reserved word in TS — the labeled arg `~class` is
+// ported as the parameter name `class_`.
 export const instanceofCond = (b: Val, class_: unknown) => (inputVar: string): string =>
   `${inputVar} instanceof ${B_embed(b, class_)}`;
 
@@ -85,6 +90,10 @@ export const int = () =>
     s.decoder = numberDecoder;
   });
 
+// PORT-NOTE: the source's `let rec inputToString = ... and stringDecoderFn =
+// ... and string = ...` mutual-recursion group falls inside this section's
+// line range, so all three are ported here (the name list in the task omitted
+// stringDecoderFn/string, but they are inseparable from inputToString).
 export const inputToString = (input: Val): Val => {
   return B_next(input, `""+${input.i}`, string());
 }
